Create game.asm
In the game, you are an "O" symbol on the screen.
Your job is to "eat" the "X" symbol.
Every time you ean an X, your score total will increase by one.
if you want to exit, press "q" key.

Note that to run the code in your local mechine, you need a DOSBOX simulator of the 8086 proccesor.